+++ authority-portal-frontend/src/app/app.module.ts
@@ -29,6 +29,7 @@ import {ProvideConnectorPageModule} from './pages/component-registration-pages/p
 import {RegisterCentralComponentPageModule} from './pages/component-registration-pages/register-central-component-page/register-central-component-page.module';
 import {RegisterConnectorPageModule} from './pages/component-registration-pages/register-connector-page/register-connector-page.module';
 import {RequestConnectorPageModule} from './pages/component-registration-pages/request-connector-page/request-connector-page.module';
+import {ControlCenterOrganizationEditPageModule} from './pages/control-center-organization-edit-page/control-center-organization-edit-page.module';
 import {ControlCenterOrganizationMemberDetailPageModule} from './pages/control-center-organization-member-detail-page/control-center-organization-member-detail-page.module';
 import {ControlCenterOrganizationMembersPageModule} from './pages/control-center-organization-members-page/control-center-organization-members-page.module';
 import {ControlCenterOrganizationProfilePageModule} from './pages/control-center-organization-profile-page/control-center-organization-profile-page.module';
@@ -80,6 +81,7 @@ import {ParticipantInviteNewUserModule} from './popups/participant-invite-new-us
     AuthorityOrganizationDetailPageModule,
     AuthorityOrganizationListPageModule,
     CentralComponentListPageModule,
+    ControlCenterOrganizationEditPageModule,
     ControlCenterOrganizationMemberDetailPageModule,
     ControlCenterOrganizationMembersPageModule,
     ControlCenterOrganizationProfilePageModule,