+++ authority-portal-frontend/src/app/pages/control-center-organization-edit-page/control-center-organization-edit-page.module.ts
@@ -1,0 +1,29 @@
+import {CommonModule} from '@angular/common';
+import {NgModule} from '@angular/core';
+import {RouterModule} from '@angular/router';
+import {NgxsModule} from '@ngxs/store';
+import {ErrorElementModule} from 'src/app/common/components/error-element/error-element.module';
+import {LoadingElementModule} from 'src/app/common/components/loading-element/loading-element.module';
+import {PipesAndDirectivesModule} from 'src/app/common/components/pipes-and-directives/pipes-and-directives.module';
+import {MaterialModule} from 'src/app/common/material/material.module';
+import {SharedModule} from 'src/app/shared/shared.module';
+import {ControlCenterOrganizationEditPageComponent} from './control-center-organization-edit-page/control-center-organization-edit-page.component';
+import {ControlCenterOrganizationEditPageStateImpl} from './state/control-center-organization-edit-page-state-impl';
+
+@NgModule({
+  declarations: [ControlCenterOrganizationEditPageComponent],
+  imports: [
+    CommonModule,
+    RouterModule,
+
+    // NGXS
+    NgxsModule.forFeature([ControlCenterOrganizationEditPageStateImpl]),
+
+    MaterialModule,
+    SharedModule,
+    PipesAndDirectivesModule,
+    LoadingElementModule,
+    ErrorElementModule,
+  ],
+})
+export class ControlCenterOrganizationEditPageModule {}