+++ authority-portal-frontend/src/app/pages/registration-pages/organization-onboard-page/organization-onboard-page/organization-onboard-page.model.ts
@@ -1,8 +1,8 @@
 import {FormControl, FormGroup, ɵFormGroupRawValue} from '@angular/forms';
 import {
-  DEFAULT_ORGANIZATION_PROFILE_FORM_MODEL,
-  OrganizationProfileFormModel,
-} from '../../../../shared/components/business/organization-profile-form/organization-profile-form-model';
+  DEFAULT_ORGANIZATION_CREATE_FORM_MODEL,
+  OrganizationCreateFormModel,
+} from '../../../../shared/components/business/organization-create-form/organization-create-form-model';
 
 export interface OnboardingUserTabFormModel {
   firstName: FormControl<string>;
@@ -25,7 +25,7 @@ export const DEFAULT_ONBOARDING_USER_TAB_FORM_VALUE: OnboardingUserTabFormValue 
   };
 
 export interface OnboardingOrganizationTabFormModel
-  extends OrganizationProfileFormModel {
+  extends OrganizationCreateFormModel {
   acceptedTos: FormControl<boolean>;
 }
 
@@ -34,7 +34,7 @@ export type OnboardingOrganizationTabFormValue =
 
 export const DEFAULT_ONBOARDING_ORGANIZATION_TAB_FORM_VALUE: OnboardingOrganizationTabFormValue =
   {
-    ...DEFAULT_ORGANIZATION_PROFILE_FORM_MODEL,
+    ...DEFAULT_ORGANIZATION_CREATE_FORM_MODEL,
     acceptedTos: false,
   };