+++ authority-portal-frontend/src/app/shared/components/business/organization-edit-form/organization-edit-form-model.ts
@@ -1,0 +1,24 @@
+import {FormControl, ɵFormGroupRawValue} from '@angular/forms';
+
+export interface OrganizationEditFormModel {
+  website: FormControl<string>;
+  businessUnit: FormControl<string>;
+  description: FormControl<string>;
+
+  mainAddress: FormControl<string>;
+
+  billingAddressSameAsMain: FormControl<boolean>;
+  billingAddress: FormControl<string>;
+
+  mainContactName: FormControl<string>;
+  mainContactPhoneNumber: FormControl<string>;
+  mainContactEmail: FormControl<string>;
+
+  technicalContactSameAsMain: FormControl<boolean>;
+  technicalContactName: FormControl<string>;
+  technicalContactPhoneNumber: FormControl<string>;
+  technicalContactEmail: FormControl<string>;
+}
+
+export type OrganizationEditFormValue =
+  ɵFormGroupRawValue<OrganizationEditFormModel>;